Move index.php home logic to libraries/entry_points
